requirements.txt
python-dotenv
aiosqlite
discord.py
requests
asyncio
sqlalchemy